+++ src/al_dvc/gui/dialogs/__init__.py
@@ -1,0 +1,1 @@
+"""Dialogs of the pyALDVC window."""